agentproject1/requirements.txt
# requirements.txt
flet==0.21.2
pyyaml==6.0.1
python-dotenv==1.0.0